fix: faker didn't work in pharo-11
Pharo 11 doesn't have the `timesCollect` method. This fixes the instances where
`timesCollect` was used and adds tests for those modified methods.

This didn't work in Pharo 11 and may not work in other versions. It looks
like `timesCollect` was added to Glamourous Toolkit, but not ported back to
Pharo.

This change also adds missing localizations for en_US.